refactor: extract pluginStoreWatcher from core
