+++ src/composables/usePluginStoreWatcher.ts
